+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Export `WeightPresetLabel` enum
 
 ### v0.41.1
 * (Demo App) Fix account tooltip when control panel isn't sticking to top by moving it into the feed `<div>`